+++ app/src/main/java/com/psiphon3/psicash/store/PsiCashStoreViewState.java
@@ -50,11 +50,11 @@ public abstract class PsiCashStoreViewState implements MviViewState {
         return psiCashModel() != null && psiCashModel().isAccount();
     }
 
-    public int uiBalance() {
+    public long uiBalance() {
         if (psiCashModel() == null) {
-            return 0;
+            return 0L;
         }
-        return (int) (Math.floor((long)
+        return (long) (Math.floor((long)
                 ((psiCashModel().reward() * 1e9 + psiCashModel().balance()) / 1e9)));
     }
